+++ src/index.ts
@@ -9,6 +9,7 @@ export {
 export { PaymentRequest, PaymentResult, AccountBalance, IlpClient } from './ILP'
 export { default as PayIdClient, PayIDClient } from './PayID/pay-id-client'
 export { default as SingleNetworkPayIdClient } from './PayID/single-network-pay-id-client'
+export { default as AllNetworksPayIdClient } from './PayID/all-networks-pay-id-client'
 export {
   PayIDErrorType,
   PayIdErrorType,